Remove KubeconfigDownload exception
