+++ src/common/components/hosts/utils.ts
@@ -1,11 +1,11 @@
+import { TFunction } from 'i18next';
+import { AxiosResponseHeaders } from 'axios';
 import filesize from 'filesize.js';
 import Fuse from 'fuse.js';
-import { Host, Cluster, Inventory } from '../../api/types';
+import { Host, Cluster, Inventory, stringToJSON } from '../../api';
 import { hostRoles, TIME_ZERO } from '../../config';
 import { DASH } from '../constants';
-import { stringToJSON } from '../../api';
 import { Validation, ValidationsInfo as HostValidationsInfo } from '../../types/hosts';
-import { TFunction } from 'i18next';
 
 export const canEnable = (clusterStatus: Cluster['status'], status: Host['status']) =>
   ['pending-for-input', 'insufficient', 'ready', 'adding-hosts'].includes(clusterStatus) &&
@@ -82,6 +82,12 @@ export const canDownloadKubeconfig = (clusterStatus: Cluster['status']) =>
   ['installing', 'finalizing', 'error', 'cancelled', 'installed', 'adding-hosts'].includes(
     clusterStatus,
   );
+
+export const getKubeconfigFileName = (headers: AxiosResponseHeaders) => {
+  const fileNameMatch =
+    headers['content-disposition'] && headers['content-disposition'].match(/filename=".*"/);
+  return fileNameMatch ? fileNameMatch[0].slice(10, -1) : 'kubeconfig';
+};
 
 export const canInstallHost = (cluster: Cluster, hostStatus: Host['status']) =>
   cluster.kind === 'AddHostsCluster' && cluster.status === 'adding-hosts' && hostStatus === 'known';
